Fix profile save, AI matching diversity, Word upload, emojis, nav
- Fix budget range validation with toInt() sanitizer before isInt()
- Diversify AI matching: type-balanced candidate pool, diversity prompt
- Add Word doc (.docx) upload support via mammoth alongside PDF
- Add success feedback screen after resume upload confirmation
- Add emojis to Market Trends headers and trend direction indicators
- Add StrategicNav to Intelligence Dashboard page
- Neutral goals placeholder to reduce category bias in matching

+++ backend/src/personalMatch/personalMatch.validation.js
@@ -15,8 +15,8 @@ const updateProfileDataValidation = [
   body('profileData.certifications').optional().isArray({ max: 20 }),
   body('profileData.certifications.*').optional().isString().trim().isLength({ max: 200 }),
   body('profileData.budgetRange').optional().isObject(),
-  body('profileData.budgetRange.min').optional().isInt({ min: 0 }),
-  body('profileData.budgetRange.max').optional().isInt({ min: 0 }),
+  body('profileData.budgetRange.min').optional().toInt().isInt({ min: 0 }),
+  body('profileData.budgetRange.max').optional().toInt().isInt({ min: 0 }),
 ];
 
 const feedbackValidation = [
